+++ src/game/CGameObject.cpp
@@ -26,6 +26,8 @@ namespace game
 		mIsAlive = false;
 		mBody->GetWorld()->DestroyBody(mBody);
 		mBody = nullptr;
+
+		mRemoveListeners.notify();
 	}
 
 	void CGameObject::onInit(IGameWorld& world)
@@ -39,18 +41,45 @@ namespace game
 		mStepListeners.notify();
 
 		for(auto& module : mModules)
-			module->onStep(*this, world);
+		{
+			// check that the object is still alive before each call,
+			// in case one of the modules decides to kill the object
+			// in this step.
+			if(isAlive())
+				module->onStep(*this, world);
+		}
 
 		// update property listeners
 		notifyAll();
 	}
 
-	void CGameObject::onImpact(IGameObject* other, const ImpactInfo& info)
+	void CGameObject::onImpact(IGameObject& other, const ImpactInfo& info)
 	{
 		for(auto& module : mModules)
 			module->onImpact(*this, other, info);
 
 		mImpactListeners.notify(other, info);
+	}
+
+	void CGameObject::dealDamage( const Damage& damage, const b2Vec2& pos, const b2Vec2& dir )
+	{
+		for(auto& module : mModules)
+			module->onDamage(*this, damage, pos, dir);
+	}
+
+	/// collision filter data. This is currently very specialised, so maybe a more general
+	/// interface would be nice. However, we need to ensure that this does not cost performance for objects
+	/// that do not require special collision handling.
+	/// Right now, we can set one specific body with which this object shall not collide.
+	const b2Body* CGameObject::ignoreCollisionTarget() const
+	{
+		return mIgnoreBody;
+	}
+
+	/// sets the body which shall be ignored upon collision checks
+	void CGameObject::setIgnoreCollisionTarget( const b2Body* ignore )
+	{
+		mIgnoreBody = ignore;
 	}
 
 	b2Vec2 CGameObject::position() const
@@ -78,10 +107,16 @@ namespace game
 		return mStepListeners.addListener( std::move(lst) );
 	}
 
-	ListenerRef CGameObject::addImpactListener( std::function<void(IGameObjectView*, const ImpactInfo&)> lst )
+	ListenerRef CGameObject::addImpactListener( std::function<void(IGameObjectView&, const ImpactInfo&)> lst )
 	{
 		return mImpactListeners.addListener( std::move(lst) );
 	}
+
+	ListenerRef CGameObject::addRemoveListener( std::function<void()> lst )
+	{
+		return mRemoveListeners.addListener( std::move(lst) );
+	}
+
 
 	long CGameObject::id() const
 	{