+++ CHANGES.md
@@ -1,3 +1,7 @@
+**Demo 3**
+ * limited lifetime of bullets
+ * first damage model
+
 **Demo 2**
  * added guns and bullets
  * added 3D graphical interface